requirements.txt
asyncio
discord.py
flask
django
requests
PyJWT
twilio
ThisApiDoesNotExist
python-dotenv
fortune
gunicorn
django-heroku
googletrans
numpy
scipy
matplotlib
h5py
tensorflow
keras
opencv-python
imageai
